New option usePackageListFiles to enable package-list support
A package-list file is a JavaDoc standard to specify the list of
packages documented under a certain JavaDoc URL.
Using the new option usePackageListFiles, you can place a
package-list file in the root of your source directory (usually
src/main/java), containing a line-by-line list of packages to be
included in the JavaDoc documentation.
This list is used to compute the corresponding sourceFileIncludes,
so that the module's local JavaDoc only includes these packages.
When building a JavaDoc aggregator (through includeDependencySources),
all package-list files of all referred modules are collected into a
new package-list file that is included in the resulting -javadoc.jar.
This again can be used with offlineLinks to generate hyperlinks
between multiple dependency-aggregated JavaDoc HTML bundles.

Switching to Java 8 (1.8) so that streams can be used.