Auto-skip in usePackageListFiles mode when package-list is empty
When usePackageListFiles mode is on and there is no (collected)
package-list file or it is empty, it does not make sense to run
JavaDoc tool. Instead of requiring every single module to explicitly
switch off JavaDoc generation in that case, we rather want to allow
configuring our maven-javadoc-plugin in a common parent POM, so that
it kicks in automatically for every module where a package-list file
exists.